Add lnr_hal
Also caught some errors in lnr_glmnet and lnr_hal:  the model.matrix.default needs to be passed a formula that doesn't refer to a y-variable, otherwise it expects that the y-variable column exists in the newdata.

Caught an error in lnr_multinomial_vglm: for some reason, it seems the VGAM::vglm function looks for "weights" in the parent-parent frame, and this was throwing errors.

Removed ≤ and ≥ from screeners.R because CRAN checks insist only ASCII code be used.

Added documentation