Improve Doubly Robust estimation article
Added a lot of expository text and references.

Also uncovered a bug (fixes implemented in following commits) in super learner for binary outcomes. The bug was revealed because super learner was picking lnr_mean over other, better learners.  This ultimately revealed to me that there was an issue in determine_weights_for_binary_outcomes().