Bugfix for binary learners
Before this, super_learner() would favor the lnr_mean over other, better learners in some examples.

Ultimately the error came from the fact that min and max were used instead of pmin and pmax (the vectorized versions). As a result, all of the probabilities predicted would get set to the min probability predicted across all values, which was not good/correct behavior.

Given this bugfix, there's also a test implemented to make sure that super_learner() for binary outcomes prefers the correct model in an example with simulated data.