Warning fix in tests- glmnet was mad about convergence
The warning messages were saying that glmnet hadn't convered for the nth lambda value.

Ultimately this was because the regression equation was too simple (am ~ hp + cyl).

Instead, just adding a little bit more information seemed to solve the problem.  Now the formula is changed to am ~ hp + cyl + mpg.  This is for a test that makes sure all the binary learners run without errors on an example.