lnr_glmnet predict error resolution; doc updates for lnr_gbm
lnr_glmnet has been a pain repeatedly, almost all related to the fact that it relies on the data being structured as an X tabular dataset and a vector Y.

Before, the hyperparameterized_learners that appear in README.Rmd were creating errors on prediction. lnr_glmnet(mtcars, mpg ~ .)(mtcars) was throwing an error that it expected 11 columns in the newdata being passed.  This is fixed by using the same approach in the prediction closure as in the model training step:  form the xdata from a model.matrix call that does not involve the lhs of the formula.